feat: Add ErrorResponse class and unit tests
This commit introduces the _ErrorResponse_ class for handling error
responses. The class includes a _responseType_, _statusCode_, _message_,
and _errors_ property. The _errors_ property is an array
that can be used for validation errors.

In addition, this commit includes Jest unit tests for the
_ErrorResponse_ class. The tests ensure that the properties
of the error response are set correctly when using the
_badRequest_ method.

- Fix #2